requirements.txt
Django	5.0.6
PyJWT	2.9.0
Unidecode	1.3.8
asgiref	3.8.1	3.8.1
certifi	2024.7.4	2024.8.30
charset-normalizer	3.3.2
distlib	0.3.8	0.3.8
django-unfold-admin	0.1.0
djangorestframework	3.15.1
djangorestframework-simplejwt	5.3.1	5.3.1
filelock	3.14.0
idna	3.8
pillow	10.3.0
pip	23.2.1
platformdirs	4.2.2
psycopg2	2.9.9
requests	2.32.3
setuptools	65.5.0
sqlparse	0.5.0
tzdata	2024.1
urllib3	2.2.2
virtualenv	20.26.2